worker/pyproject.toml
[project]
name = "docking-worker"
version = "0.1.0"
description = "Celery worker for docking pipeline"
requires-python = ">=3.10,<3.12"
dependencies = [
  "sqlalchemy==2.0.32",
  "psycopg[binary]==3.2.1",
  "celery==5.4.0",
  "redis==5.0.8",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "rdkit-pypi==2022.9.5"
]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
